Bulk Commit. Splitting up mathMagical I split up the mathmagical code into diffrent files. I decided that deck.java wasnt needed and so deleted it. The deck method in now in cardFunctions. This was made much easier by VSCode since it has a capability to move over methods to diffrent files.
The calling a new deck is now difffent that how it used to be so some
alterations to the cardValue class was needed. Mainly using .equals
instead of ==. All other changes have been just moving exisitng code.

I have a launch.json file so its easier to change what cards are passed
in as arguments for the moment. However my tenative goal is to have the volunteer
program takin in argument for the program interactivly.