Interactivity Implemented I made alterations cardFunction methods that removed the need for the deck() to be passed in as a variable. This seems superflous.
Volunteer now will take in card values from user and verify that they
are acceptable card values

magician was altered to take out the need for deck()

mathMagical will now be the main hub for all the code to be called.

assistant now returns a String[] so it can be passed along to magician
within the mathMagical file.